Check markdown rules in github actions
This way we'll notice syntax inconsistencies in pull requests. We
weren't doing it until now because there's no Pronto runner for MDL and
we weren't running linters individually.

Note we aren't checking the templates in the `.github` folders, since
these templates don't have a top level header in the beginning (they'll
be displayed on a page which already has a top level header) and MDL
doesn't provide a way to ignore certain rules on certain files.

We're ignoring the files included in node packages. We could also ignore
the files in node packages using `mdl -g .`, which ignores files that
aren't under version control. However, that would also check the
`.github` folders.

Finally, we're ignoring the `vendor/bundle/` folder, since that's where
gem dependencies are installed when running github actions.